chore(daemon): split out hook and server impls
