+++ app/daemon/src/gateway/handle.rs
